Spraylist integration
I seriously hope this is the last time I will have to touch anything
remotely related to spraylist code.

Size of preallocated memory, use all available allocators, transactional
adds.

+++ lib/spraylist/skiplist.h
@@ -36,6 +36,7 @@
 #include "atomic_ops_if.h"
 
 #include "tm.h"
+#include "utils.h"
 #include "ssalloc.h"
 
 #define DEFAULT_DURATION                1000
@@ -59,7 +60,7 @@ extern uint8_t levelmax[64];
 
 #define TRANSACTIONAL                   d->unit_tx
 
-typedef unsigned long val_t;
+typedef uint32_t val_t;
 typedef intptr_t level_t;
 #define VAL_MIN                         INT_MIN
 #define VAL_MAX                         INT_MAX
